GDA-5480 Add a --client command line option to dls-config/bin/gdaclient
Fix the bug where a different parameter order would cause different
behaviour.

Fix the mishmash of tab and space indents.

Rationalise the three similar sets of help text down to a single help
text function.

Add a --keep option which causes a workspace to be renamed with
todays date and time rather than deleted, allowing subsequent
root cause analysis.

Add a --log option which disables the automatic start of the log
client.

Add a --nohup option which causes the client to start using nohup,
with output being redirected to the specified file.