DATAACQTEAM-62 - Add a standard command for starting Eclipse on a beamline.
This bleclipse command needs to be location independent so will work
out which workspace to open depending on the location of the script.

It unloads modules which shouldn't be loaded, loads modules which should,
adds standard ECLIPSE_OPTIONS and populates the ECLIPSE_KEYRING with the
standard location if ECLIPSE_KEYRING isn't already defined.