BLX-18 BLX-19 (8.30) Control first energy scannable during CV Energy Scan
BLX-18 Added better tracing to ContinuousEnergyMoveController and the
ability to control a single energy scannable. Also added the ability
to provide a PositionCallable which returns the position of this energy
scannable at the nominal (dead reckoned) scan points.

BLX-19 Added better tracing to ContinuousEnergyScannable and updated
ContinuousEnergyScannable to get it's PositionCallable from it's
ContinuousEnergyMoveController.