BLX-50 (8.30) Fix problems with improved POMS support found in testing.
Note that for some reason GDA's jython interpreter does not like
evaluating calcs with parentheses in, so the FlipperCalcDeviceClass
is far less useful that it could be.

Since the output calculations are always performed in order this problem
can be worked around by using an output as an intermediate value for a
later output. 

Q1 is always evaluated before Q2 etc. this can be worked around
by using one output