BLXXJ-40, BLXXJ-32 change to Strip detector interface to connect after configure
Now StripDetectors should not connect to hardware during their
configure, but within a connect() command.

The UI will work out which one is connected via the ClientConfig object.

Still some work to do on this.

I have also added static methods to EdeScanParameters as a convenience
to make simple scans for StripDetectors. This is for BLXXJ-32