BLXXJ-65 refactor the structure of the scan classes and add a new Single Spectrum driver class
The classes which were all in gda.scan are now in a heirachy to
hopefully make their purpose easier.

I have moved code which was in Jython into Java which drives the single
spectrum experiment using just native types. This has enabled it to be
unit tested in a similar manner to the actual EDE scan classes.  This
class should be used within Jython scripts or from UI code to run a
single spectrum experiment.