BLXXJ-20 copy the D10 camera configuration from I20 into I20-1
At the moment I20-1 does not have its own x-ray eye camera. As a
temporary measure the D10 diagnostic camera from I20 is included in the
EDE configuration as the current ways of doing things is to move the D10
camera over from I20 into EDE and plug it into the network which is OK
as the network is shared.