BLXXJ-137 linear experiment now broadcasts latest spectra and progress to a distributed ScriptController object to be picked up by UI
EdeScan object now can be given an IObserver object to sent every
spectra to. This is only used by the LinearScan to get updates from the
itScan - it then normalises each spectra and in turn broadcasts out the
normalised spectrum to a ScriptController object it gets from the
finder.

Maths for normalisation had to be moved into static methods so that the
linear experiment could do normalisation on-the-fly.