BLXXJ-138 add an optional reference sample position to ede linear scans
This takes a single spectrum for each timing group just after the
initial I0 using the given IRef position.

An ascii file is created with the derived data based on the initial I0.
I don't see a need to wait until the end of the experiment as that I0
should be correct as it was taken immediately before the IRef data.