BLX-45 (8.34) Add GDA_CLIENT_WORKING_DIR option to support relative configs.
This option allows you to specify the working directory from which the GDA
client is run.

This can be used to specify a directory which is not writable, and thus
effectively disable core dumps, or it could allow relative relative
directories to be used in client .ini files (e.g. use ../../var to get to gda_versions/var).