BLXV-82 (8.36) Complete PE support for expose.
Re-enable HDF5 writing as it is now working and add the lazyOpen property.

Add beans for the other plugins defined in EPICS.

Add support for a MultipleExposureSoftwareTriggerAutoMode collection
strategy and switch filewriters to force their input array port to be
the output of that process.

Also remove the old pe and peAD detectors along with the pedet_for_snaps
bean and rename pedet to pe. Remove obsolete contents of PE dummy device.
Remove pe_adbase and use peAD_config_base instead.

Rework which beans are internal to another bean and which can be pulled
out and made common between beens (such as pe_UnixToWindowsFilePathConverter).