BLXV (8.36) Convert logging system to remive duplication, using includes.
Note that there is still a problem with this method in that some paths
have to be hard coded due to problems with relative paths.

Paths relative to the client seem to be fine, but the ones for the server
seem to be failing, so I don't understand what the working directory of
the server is being set to.

Also added debug=true to all config files.