BLXXJ-264 each EdeScan can now test for top-up before it starts its data collection
If it does and how long to leave before the next top-up is given to the scan by the outlying Experiment object.

This gives a degree of control of the flow to the experiment to the Experiment, but the actual work of testing for top-up is performed inside the running scan.

This has required a change to the EdeTimeEstimate interface to make the values it retursn more flexible. This will enable the extra logic for the external TFG in the TimeResolved and Cyclic experiments.