MTGDA-173 (8.40) - Add gda_setup_env to b16-config/bin
The Swing client assumes there is a gda_setup_env in the beamline bin
directory. It didn't matter that it didn't exist since the beamline
workstations had the required environment already setup
(i.e. $JAVA_HOME was already set). This was changed and so this file was
required to exist.