XAS-190 remove GDA copy of RichBean classes and use DawnSci version instead
Changes in GDA dev since the fork ahve been applied to the DawnSci version

Change-Id: I65d52fa37bf48f3c94e7ebad8e6fe5f78be52b8c

Rename IRichBean interface to XMLRichBean

This interface is a marker for bean classes which represent something which will be displayed in an UI, via the BeanUI class, and persisted in XML via the XMLHelpers class.  The BeanUI class uses Object in its signatures so XMLRichBean is only for GDA classes wanting the XML persistence.

Change-Id: I37bc26dbfae900ad797e70810a710f0bb79f1a4f