BLX-7 (8.44) Add a zie FollowerScannable and improve logging in FollowerThread.
The Zero Input & Extra name scannable SilentFollowerScannable can be
used inside a cvscan to turn on the follower behaviour at the start of
the scan, and off at the end.

This is simpler than implementing ContinuouslyScannableViaController on
FollowerScannable, since the only values we want to monitor will be
provided by other CSVC scannables any way.