BLXV (8.44) Add a zebraPositionScannable to check for zebra vs epics position
Since it is possible for the position zebra believes the motor to be in
to be different from the position Epics thinks it is in, we need a way
to compare the two.

Zebra doesn't allow us to read out it's internal position directly, so
instead this scannable causes a position compare to be performed whenever
it is moved.

This scannable should not be used at the same time as any other zebra
position compare scannable, as they will conflict with each other, as
discovered when attempting to create a ZIE (zero input and extra name)
scannable which could be used as a default scannable for checking
that the zebra position matches the epics position at the start of each
scan.

This commit also contains an abortiver attempt