BLXV (8.44) Move <logger> configs into an inluded common.xml file in dls-config.
This allows logging properties to be kept consistent between sources, and
sinks. Without this, a mismatched property could be sent from the object
server to the log server, but then be discarded and not forwarded to the
log file or log client. Similarly, a message might be allowed theough the
log server, but culled at the source side.