Changes to allow i13i servers to start in dummy mode.
The majority of the changes consist of adding dummy objects to parallel
the real ones, but note:

- Removed launchers to do with i12

- The shared version of d1.xml (in i13-config/devices) was missing
and had to be copied from 8.40

- Most of the contents of devices/live/p2r.xml moved to
new file devices/p2r.xml

- devices/pco1.xml moved to live directory: dummy equivalent created.

- Property gda.beamline.scripts.dir used to be defined in dls-config
but is no longer, so has been put into the appropriate
java.properties in i13-config.

- localStation.py modified so it does not start a number of devices
in dummy mode.

- includes of devices/float*.xml moved to common.xml

- dummy server.xml modified to be much more similar to
live_server_common