Add a dummy-mode product definition and synchronise the launchers
Now the launchers are kept only for convenience. It should always be possible
to run the live or dummy client from the product file overview page, and this
ensures that the client application used in testing is as close as possible to
the client that will actually be built. The two product files should be kept
as similar as possible to help with this.

The offline product file includes variable substitutions in the VM arguments.
These will work when launching from Eclipse but not when the client is built.
If the built client is run using the standard GDA startup scripts all of the
necessary VM args should be set anyway.

Change-Id: I9b5bb28648c7dec77405d50bf27b62d360b51137