First draft of script to degas slits
This replaces the previous degas script, which seemed to be incomplete
and not what is required for i21.

This version is a first draft for running in simulation, with sleep()
calls to slow it down so you can see what is going on. Before running
it on the beamline, we must at least:

- remove many of the sleep() calls
- put in more accurate defaults for the pressures at which the shutter
should be opened/closed
- check that the the positions given for slits to be open/closed are correct