Change dummy front end to behave more like the real one
The real front end shutter reports its position as 'Closed' when shut,
but the command the GDA needs to send is 'Close'.

The dummy object does not make this distinction, so I have changed
the value to 'Close'. This means that the degassing script will now
function correctly, but the output of the pos command will not exactly
match that of the real object.