Move i05-1 classes that map to VGScienta devices into VGScienta device package
These three classes and their associated Spring mappings relate to VGScienta device classes. The Fake also has a duplicate in the i05 beamline plugin which is referenced by the i05 config. Moving them here removes the need for the server to depend on the uk.ac.gda.beamline.i05-1 package which contains mostly UI related classes. The moved classes are only referenced by the rest of the i05-1 config.

Change-Id: I0f110802c0ebea2f82eef2108e713bb2f186d522