Prepare for uk.ac.diamond.daq.dls.controls and other i18 bundles
The gda.scannable classes currently in uk.ac.gda.dls are going to be 
moved into a new uk.ac.diamond.daq.dls.controls plugin to break some server to 
client dependencies. This change prepares the way for that and also 
comments out references to a logging class that doesn't exist any more.
The new dependencies are currently marked as optional since the plugins 
that they refer to are not checked in yet. Additionally it adds links 
to allow the i18 config to appear to the gda server as if it complied 
with the 1.0 layout standard.

Change-Id: I20c88f903ce570302208bba4b2d4e9a1095ee9eb