BLX-91 (8.50) Fix issue with cvscan hanging.
Allow a fixed value to be added to the struck scaler total acquire time.

Sometimes the RASOR struck scaler returns 1 fewer points than requested,
this causes the binpoints to fail and the whole scan to fail. Adding the
shortest possible time to the total count time seems to ensure that all
of the required points are acquired.