I20-4, I20-5 : Implementation of fast shutter during scan, add GUI controls for topup
A checkbox has been added to the GUI to control whether the fast shutter is used during a scan; when using main shutter only, it opens and closes at the same points in the scan as when using the fast shutter.
A checkbox has been added to the GUI (Linear experiment view only) to allow the topup checker to be switched on and off for main It collection.

Change-Id: I4029addbae421db01f3fa423390820d5f3639262