I20-9 : Crash causes client to lose the experiment mode
XES/XAS experiment mode preference is now saved to disk immediately after selection by the user on first startup of the client. Previously this was only written when the client exited (cleanly). This fix helps to avoids the problem of getting stuck in the wrong experiment mode if the clients crashes before the preference can be written to disk.

Change-Id: Ia14255d91c5b7c33d24ce61d8584944cc9067dc0