Add scripts to read/rewrite server target links on demand
As the GDA9 server is primarily a standalone built application, there 
are a number of possibilities for which version will need to be started 
on a beamline when scientists use the gda servers launcher (equivalent 
to running the gdaservers script with no parameters). These consist of 
the current released version, the latest snapshot build, the local
development build or the one located at a specific path. The latch 
script reads or rewrites the new 'launch' link in the beamline's gda 
directory on dls_sw which is the default start target for 
remotestartupscript.sh. Using this the user or other scripts can 
re-point the link as required to latch the version that will be started 
in future.

The read_release_version script is used to determine the released 
version that the beamline workspace was materialised from so that this 
can be used to ensure that the 'release' target set in the latch script 
is version bound, i.e. doesn't change when a new release is made. At 
the moment this is determined from the pewma-script.txt file written by 
pewma when the workspace is materialised. In future this information 
will hopefully be available from an eclipse materialised workspace too.

Change-Id: Ic82c6bc143c46ebb1c54acf15e7bf531e052c393
(cherry picked from commit 7d36c9ffa16440aef09760dce75d5541e5e0d074)