Coherence: Improved shutter handling in dummy mode
In the client: move the Epics definitions of expt_sht_stat &
fast_shtr_stat to live client and introduce dummy versions.
This speeds up the client startup, as it does not try (and fail)
to communicate with Epics.

In the server: modify dummy eh_shtr so it has the same positions as
the live object. This removes the error messages when clicking
the Open/Close button.