Qexafs configuration for dummy
Conflicts:
	i18/servers/main/_common/detectors/qexafs.xml
	i18/servers/main/dummy/detectors/qexafs.xml
	i18/servers/main/live/detectors/qexafs.xml
	i18/servers/main/dummy/detectors/xspress3.xml

Effectively a cherry-pick of commit ea847fe69dae87b7ca0b9a09b954dd98819943e0,
but done manually so the changes have been applied to the right parts of
the files in their new locations. Most of the work of the original commit
ea847fe6 on gda-9.0 was duplicated by commit cd318841 on master so this
commit just makes some small adjustments to make sure the master config
duplicates the function of the current gda-9.0 config on the beamline.

Change-Id: I88ee51c3ca0558bc4dbff29d68b27d45f7caf60a