I13-143 Coherence: move customisation to the i13j plugin project, so it read by default
Previously, most of the customisation was in pluginCustomization.ini
in the config project, but this relied on its location being passed
as a VM argument, which in general the GDA startup scripts will not do.

However, anything put in plugin_customization.ini in the relevant
plugin project (in this case uk.ac.gda.beamline.i13j) seems to be read
by the Eclipse framework without specifying any VM argument.