DAQ-228 Ensure new logs directories are world-writable
After testing on I18, it seems the "mkdir -m 777" doesn't actually
create a directory with 777 (a=rwx) permissions - the directory ends
up with 775 permissions instead. Trying to start the servers as i18user,
that means the directory is owned by i18user and not writable by gda2
and therefore server startup fails. Explicitly setting the permissions
to 777 using chmod seems to work (I also tried to make it work using
umask but couldn't).

Change-Id: I36ddb8b7824f24cec6dbc2d38cff701e96468839