I05-100 Create a analyser object which gets the data directly from CAM
This should reduce the ~250ms delay which seems to be caused by waiting
for the frames to reach the Linux IOC over the EPICS v4 bridge see
I05-94. This design is more like the approach used on I09

This detector will not support changing of the number of iterations.

This add a dummy device for testing the live config will be in another
commit.

Also:
- Optimised summing the ROI
- Added performance logging to get more understanding during real use.

This has now been tested and used extensively for one week of user
beamtime. The performance was as expected and no problems were observed.

Change-Id: Ifeafd759d6737bd5eb3f5c30b333f59f73d95f36