I05-80 Disable always delay to allow the HV to be ramped when needed
This only changes the setting to not always delay i.e when no voltages
need changing the delay will not be used. The region delay need to be
set in SES for this to work. This change should be reverted once DAQ-259
is completed.

Also changed to use caput_wait with 1.0 sec timeout

Change-Id: Id9aa85c8653854e74c9bd116cc16596f3f058e9b