DAQ-196: Handle startup errors written to server startup file.
As a result of http://gerrit.diamond.ac.uk:8080/#/c/4398/, failure to 
start the server from the standard scripts will result in an accompanying
error message being written to the object_server_startup_main file which
is monitored by lookforfile. This change detects the presence of such a 
message (the file will be zero length if written by the application 
itself) and exits displaying the message using the bashlog colour scheme.

Change-Id: I48dbe40d39dc6611bc29dd503b24604412876a77