I09-15 Fix GDA requesting too much data from the analyser
There were 2 causes of this issue. The first appears to be a bug in JCA
where if a monitor is added to a PV the full size of the PV will be sent
in every update even when only a small number of the elements are
relevant. The second is that when a small number of frames are used the
rate of updates is high even though that rate is not required for live
visualisation.

This fix changes a few things:
- The client now monitors a PV which informs it a update is available
  rather than the large array PV. BL09I-EA-DET-01:CAM:CURRENT_POINT_RBV
- If an update is available a rate limiter is used to decide if to get
  that update, which can be configured in Spring.
- The actual data is requested by the view which needs it over RMI with
  the specific number of elements.

Change-Id: Ifb642eef5a148deafa49f6fe91913069b0bd7c88