I20-50 : Updates for setting Medipix/Merlin ROI from stored values in XML
The Medipix ROI is now set from values in XML file at start of the scan. After the scan, the ROI provider used in the ADRoiStatsPair plugin is changed back to its original value (i.e. plot server).
Also, the Metashop object is now passed to output preparer so that the metadata is correctly written.
Unit tests have been updated to include medipix and metashop.

Change-Id: Ic825e9267f4819726c20b65dfa34373e014c7ae2