Add required uk.ac.gda.devices.vgscienta plugin to i09 and i09-2 features
This is required because electronanalyser.server is currently required
in the client. Ideally this dependency could be untangled and removed
and this would also remove this dependency.

i09-2 is currently undergoing rebuilding and will require significant
changes to the client for the new endstation which will involve removing
these dependencies as they are changing to a different electron
analyser.