I05-105 Change GUI to use new RMI interface and remove hard coding
This makes some significant changes to the GUI for controlling the
analyser on i05 and i05-1.

Essentially it is switching from using AnalyserCapabilites object to
know about the analyser to using a new VGScientaEnergyRanges object.
There is also a new interface so the server side analyser object can be
directly interacted with by the client. The idea is that after a little
bit more cleaning up all the analyser logic will be server side.

The AnalyserCapabilites objects held a lot of hard coded data about the
axis for different lens modes which we can now get dynamically from 
EPICS, and from VGScientaEnergyRanges object.

Change-Id: I1b2f96b32c0d5897056a0b6c08086e35e306dbad