Update i18 config to refere to moved microfocus scripts
Currently the scripts in uk.ac.gda.client.microfocus are copied to the 
dls-root folder in the exported product to make them available to Jython. 
This plugin is not part of the product and not suitable for the new 
extension point mechanism. 

To get round this, the scripts and the ScanPositionsTwoWay class are 
being moved to the uk.ac.diamond.daq.microfocus.scan bundle, this change 
updates the i18 config is accordingly

Change-Id: I353b3a3114ac45a9bf81711b6621f8f2b1fbda73