B07-40 Add ROIs and cameras for d1c, d2c, d3c, d4c, d5c d6c and d8c
This change involves adding the items to the menu and the views via
extension point. The secondary ID of the view is used for the server
side to capture the ROI via the plotting system so must match. Also
added the ConfigureAcquireTimeDecorator so the exposure is set when the
scan starts. All the cameras support 2 ROIs

Change-Id: I7f92203d7c94bcdb28f61e7ac05f86a9ecce5f65