Include Scanning command scripts location in standard locations
Switch the reference to the Scanning command scripts to use the
extension point rather than the old style script project so that it is
available to all beamlines in GDA-9. Doing so in core means that the
org.eclipse.scanning.command plugin is not required to depend on the
plugin that defines the extension point.

Change-Id: Ia39676fca18221f6049af8c43e942cbdc8d5dbd5