I14-88 Configuration for CAEN BEST
Add i14-best01 to com.cosylab.epics.caj.CAJContext.addr_list

This is required in order to control the CAEN BEST, a beamline
stabilisation device which hosts a vendor supplied EPICS IOC
for control and monitoring.
Since it runs an unmanaged Ubuntu installation, IT have
requested it be moved to the instrumentation network.
However this means the PVs served by the IOC will not be
visible to Channel Access clients on the beamline network by default
– the above environment variable setting should fix this.

Change-Id: I2155edf5e1de21fb81d2bc5e81a83ef8776f29af