I15-207 (9.1) Implement full PerpendicularSampleMotion.py algorithm
Previously the PerpendicularSampleMotion.py only calculated changes to
dx & dy for changes to phi, assuming that the phi axis was horizontal,
with specific values of kappa and theta.

It now needs to support arbitrary phi, kappa and theta positions, 
calculating dx, dy and dz.

Since the test cases for the original algorithm failed with the new
algorithm, the original behaviour has been left as the default, and the failing test cases have been disabled.