I20-129 : Turbo XAS data collection improvements
Data collection is now done in a background thread whilst the scan is taking 
place. It monitors the number of frames available, and collects a new spectrum
when enough new frames are available. 
The data collected is now determined by the available data in the buffered 
scaler used for the scan.
This commit also fixes the Nexus writing that was previously broken due to 
the default value for NexusGroupData.isDetectorEntryData being changed 
from true to false.
 

Change-Id: I2b8573fc4053c57353058d42ea8134b35d3869f3