I20-37 : Time resolved experiment It collection for different scannable positions
A scannable to be moved during a scan along with a list of positions to use 
has been added to EdeScanMotorPositions. This is used in EdeScan to allow 
the whole data collection (typically just It) to be repeated several times, 
once at each position of the scannable. File processing has been updated 
accordingly (and tidied up); unit tests added to make sure data is stored 
and processed correctly. 

Change-Id: Ieb57a14e306a7fdc5274792dfd1e0212b1b8272c