I15-207 (9.1) Disable dperp/dpara/dheight as they will fail once axes are fixed.
These scannables are not currently needed, and by the time they are next needed
it is likely that one of the axes will have been inverted and these calculations
will be invalid. This is a reminder that this code will need to be revisited and
tested the next time it is needed.