I20-143 : Disable timing group GUI controls when scan is taking place
Timing group controls are now disabled (i.e. greyed out and non selectable) 
if a scan is taking place. Since these controls communicate with the frelon/Xh
detectors to find number of accumulations etc, this prevents the user accidentally
interfering with a running scan by making a change in the GUI to the timing group setup.

Change-Id: I677825a41b4fa57af7e5e41823716a9faa4c05d8