I20-146 : Implement circular buffer for scaler data during TurboXAS data collection
When setting up for a scan, instead of trying to store all frames for all 
spectra in scaler memory at once, we instead use 'cycles' feature of the Tfg
and set the number of frames per cycle to record a whole number of spectra. 
During collection, once each spectrum has been read out (and stored in the 
Nexus file), the scaler memory locations for that spectrum are then cleared, 
ready for use on the next cycle(s).

Change-Id: I8d2dbf5286be67bfa82a654a5603f8dff82e4d9b