Stop (hopefully) ssh hanging in live mode
People have observed the ssh connection used in live mode not returning
properly in the past which is why it was backgrounded. According to
StackOverflow this can happen if the program it starts leaves its std
streams open and the recommended workround is to redirect them all to
/dev/null, which this change does. This allows the ssh tunnel to be run
without the & and thus removes the need for spurious 'read' lines in
beamline specific scripts to prevent the tunnel being killed before the
server has had time to start. This supports change 6110 and change 6111.