Move FakeNDPluginBaseForSwept to a shared server side package
This class is needed by i05 and i05-1 the server side class was in a
i05-1 specific package even though its not i05-1 specific so it has been
moved into the general vgscienta package. The copy which was in
uk.ac.gda.beamline.i05.analyser has been deleted and the spring
references have been updated.

Change-Id: If4bb66d71c956816c484eeed6c00a7ec2531f1fb