I05-148 Make a configurable entrance slit class
This is a refactor of existing beamline specific classes to make this
more configurable. It now allows the PV and slits to be configured in
Spring and monitors EPICS for changes of the slit. Also created an
improved dummy class which has the same Spring as the real class but
with the class name prefixed by "Dummy". Update the i05 and i05-1
configs to use the new classes.

It creates a new static nested immutable class to encapsulate the slit
data.

Also
- Rename labelPv to basePv, to follow a weak convention
- Replace positions Vector by List
- Remove unused label from classes and interface
- Add comments and JavaDocs
- Add output format to EpicsScannable (only used for terminal)

Change-Id: I194822589bfbbea2fefe95d2fe28e022a822b1b0