Make use of centralised remote startup script
A template of how to use the centralised remote startup script added to
dls-config allowing custom commands to first be executed. Could easily
be moved into /config/etc/live if that were to become the standard
location for these, but that would require update to the authorized_keys written by CFEngine plus adjustment of the relative path.

Change-Id: Ib3de6821fd13cb6093defdf8b5353502007911fd