Provide centralised remote startup remote startup script
The majority of non-mx remote startup scripts are functionally identical
(in some cases they are completely identical) so it makes sense to have
a centralised script that the beamline ones can just source. This will
reduce duplicated code whilst still allowing the facility to execute
custom commands before invoking the shared script.

The script determines the location of the correct gda script to use
based on the presence of the config link which is now a mandatory part
of a deployment in GDA9 removing the need to know which beamline is
being started. It could be invoked directly from the control machine's
authorized_keys file, though this would remove the ability to customise.

Change-Id: Ib55e8a7fd8699de1d757508df7c4a8306d9b65e8