I07-69 - add stats for full pilatus images
Apply the stats plugin to the output of proc (which is applied to CAM
output).

Does not use the normal ADRoiStatsPair stuff (instantiating an
ADTimeSeriesStatsPlugin directly instead) since we don't actually want
to filter the output through a ROI plugin. The Stats plugin in GDA still
expects a ROI provider, so one must be provided.