I07-73 - reconfigure fastpil output directories
Write output images to where detector images have always been written on
I07. This isn't necessarily "standard", but I don't think it matters.

I have some concerns with writing to the same directories as the old
ADPilatus system, as Eric's code does not manage the file-numbers in the
same way and so there is risk of overlaps, but images written under that
system are prefixed (e.g. with "pil100k"), so no data should be
overwritten.