Use absolute path in calculating base for relative paths
Depending on what is in the authorized_keys file on the control machine,
the remote startup script can end up calculating its location path to
include symlinks (like/gda/bin). This results in a non standard base
location from beamline to beamline from which to determine the relative
path to dls-config. To prevent this use the -P option of the pwd command
in the calculation so that the absolute path of the base location is
always returned. Update the gda script as well for consistency.

Change-Id: I8877efc15e05c8633950f47edb6da9762b504398