I07-70 - add means to create and apply masks
Adds configuration for the AD Python plugin to fastpil detectors,
and provides a AdPythonPlugin class for the IOC to use in the plugin
which can apply a mask created by the Jython class ADMaskWrapper
(stored in /dls_sw/i07/var).

It's a little clunky, but this completely replaces the previous use
of the proc plugin and could prove very useful later on.