I08-47 Update I08 product files
The result is a single product file, geared towards running the client
in dummy mode from Eclipse. The fact that many of the properties
point to dummy-mode files doesn't matter, as these will be overridden by gda_client_core when starting the client from a script.

The stated task I08-47 has been done as part of this by removing
the property:
-Duk.ac.diamond.scisoft.analysis.analysisrpcserverprovider.disable=true

Change-Id: I57047f9c0cc9501b11feda62b8ac17c5b81ecf7d