I15-1 (9.4pre) Fix i15-1 live product.
Revert live product id to uk.ac.gda.beamline.i15-1.product since this is
used to generate the id of the feature it uses and there is no separate
i15-1 live feature.

Then then add a uid to distinguish the two products, so that Eclipse can
work out which product file to use when building the i15-1 product.

Confusingly, the "Product" field in the Eclipse product editor Overview
corresponds to the "id" the xml, while the "ID" field in the editor
corresponds to the "uid" the xml. The "uid" is only in the xml if the
field is populated in the Eclipse product editor, i.e. it is different to
the "id" field.