I20-162 : Setup gui from scan parameters; save scan parameters from gui
Buttons have been added to the Linear and Cyclic experiment GUI views to 
allow the current settings to be saved to an XML file, and to load a previously 
saved XML file and setup the gui appropriately. XML files are serialized i
nstances of TimeResolvedExperimentParameters object.
The start and stop scan buttons have also been moved to the bottom of the
views.
Some minor refactoring and tidying up (removal of old commented out code) was also done.

Change-Id: I80b9db87f77375d2bb012fcd3b06ef2731d8053d