Don't use lookForFile to determine if the server has started any more.
Use of lookForFile to determine if the GDA server has started is now
deprecated and unnecessary as the GDA_SERVERS_UP environment variable
is now set when the telnet ping on the server indicates successful
startup. With the revised remotestartupscript, this is now always
determined before execution returns to the gda script and so no extra
waiting of the form previously provided by lookForFile is necessary.

This change also prevents the user being asked if they want start the
client when they have just stopped the server or the gda script never
returning because lookForFile is looking for a file that has just been
deleted.

Change-Id: Id3c6616de20fd750ba6d612c69514b26553aca64