I15_1 (9.4pre) Set org.dawnsci.mapping.ui.processing.off=true for GDA client
When debugging detectors, without this set to true the log can fill
with DataServer messages. With it set, the default Mapping and Scanning
Perspective visualisations aren't run, but live on the beamline the new
DataVis Perspective is far more likely to be used in any case.

See http://confluence.diamond.ac.uk/x/UaM-Aw for more details.