I20-189 : Improvements to topup marking for Tfg triggered Frelon scans
Scaler data used to mark spectra recorded during topup is now collected 
during the scan. Previously, the scaler data was all read at the end of a 
scan, which could lead to a large pause at the end whilst the data was 
read from the daserver.
Frame data is now read from the scalers only during and close to topup, to 
help reduce potential slowdown of data collection; zeros are used for the 
topup marks for all other times.
The topup mark for each spectra is stored in the nexus file in the dataset
'scaler_counts'. Use of real scaler values for the topup value is shown by the dataset 'is_topup_measured_from_scaler'

Change-Id: Ic04cbbe9e008b702bcc3751d8d31d215427e7c73