I20-198 : Make calibration tool work for TurboXas scan data
The Ede calibration tool will now also load data to be calibrated from a
Nexus file. When loading from a Nexus file it will look for a dataset called
either 'pixel', 'strip' or 'position' to use for the x-axis values and so should
work for data collected using single spectrum scans (Frelon or XStrip/XH
detectors), as well as TurboXas scans.

Change-Id: I3a8af21735826ea0d47e338101e1109546f5103a