I16-55 - Use moveTo in code for th2th
This is a workaround, not a permanent solution.
asynchronousMoveTo(..) will trigger the deferred move mechanism, so
calling it on another scannable within the same group will try to
setup a second deferred move on the same group, which will fail.

However, waiting for each motor in turn to finish defeats the point
of th2th, so this needs a better fix.