I15-156 (9.4) Open fast shutter during detector shield open check
Previously, the diode check would simply check the value of the diode.
Since the fast shutter was typically closed at this point, it didn't
actually check whether the beamstop was in the correct place, as there
was no beam.

Instead, the fast shutter should be opened briefly, then the diode checked
and finally the fast shutter should be closed, before the detector shield
can be opened.

Change-Id: I22d4306974f1c1931550752a753208a17ac8d05c