DAQ-690 Preserve old logging settings so they can potentially be re-used
The Server config has been set to use the standard GDA 9 logging setup
for now, however it may be desirable to reuse some of the previous
formatting etc. This change just renames the logging folder rather than
deleting it to allow for this possibility. 

If the default setup is acceptable then it will be removed at a later
stage.