I05-198 Fix copying of initial .arpes file to scripts/visit dir
Realised this was being done in a perspective listener which wansn't
ideal as that code is run every time the perspective is changed so the
code was moved into the creating the perspective.

The reference .arpes file is now stored in the config root with the file
name specified by the property gda.analyser.sampleConf

Also refactored out the analyserListener which zero's supplies when the
perspective is changed and no scan is running.

Change-Id: I2346915a41c59e0b842bf503f25652e1491cc60e