I20-215 : Allow dummy data to be use for parts of Ede scans
This change allows data to be loaded from a file and used in place of real
detector data. E.g. to substitute data for the light I0 part of an Ede scan 
with values from a previous measurement.
 - New EdeDummyDector - readout can auto generate data, or return values 
from a dataset (which may be loaded from a Nexus file).
 - TimeResolved/SingleSpectrum scans can use specific EdeDetector for parts of 
the scan with a particular combination of position (inbeam, outbeam, ...) and 
type (light, dark) 

Change-Id: I9ababa7e58416d8d098052addc386f42ebb95691