I20-226 : Create ascii files from Nexus files produced during EDE time resolved scans
 - A new static function has been added ('TimeResolvedDataFileHelper.createAsciiFiles')
to allow the Nexus files produced during time resolved/linear experiments
to be re-processed to generate the ascii files.
 - A new flag, 'writeAsciiData', has been added to TimeResolvedExperiment
to switch ascii file writing on and off.
 - Unit tests have been updated to check that ascii files are created after
scans and that they have the correct number of columns and rows; A unit
test has been added to check that the new 'createAsciiFiles' function works correctly.

Change-Id: I42162a2db3cab2bf6179936f5c2ca1786cc8d775