DAQ-834 Update beamline start scripts so that --force will work
The gdaservers was still in the form that was use when the targetselect
script was active, but the 'source' for it was commented out (as the
script no longer exists). This meant that the $ARGS and $operation 
parameters in the last line would not have been and thus --force was not
passed to the underlying gda script.

The contents have been replaced with the standard form which just passes
its parameters straight through.