I20-247 : Add a scannable/PV to be recorded for each spectra in a linear experiment
This change allows the value of one or more scannables to be recorded for
each spectrum during a linear Ede experiment. 
The PVs may also be used to provide values, in which case a simple scannable
is created with a getPosition method that returns the result of caput on the PV.
The list of scannables/PVs to be monitored is serialized as a Map<String,String>
which is used to locate/generate scannables at scan run time.

Change-Id: Ia2c412d2442c331aad18b0e81693dfc83cce9994