DAQ-774 Add units to the Mapping UI
This adds unit conversion for relevant regions and paths in the UI side.
The models remain always in mm as is currently required for Malcom.

It is implemented using a new NumberAndUnitsComposite which handles
doing the conversion for mm, µm and nm. Using this composite requires
the auto-generation to be dropped. This is handled by classes managing
which regions or paths have custom composites. If they exist they are
used otherwise fall back to auto-generation.

There is a property to set the default units per beamline
"uk.ac.diamond.daq.mapping.ui.deaultUnits" which can be mm, um, or nm

Known limitations:
- Only works when models are in mm (already a Malcom limitation)
- Only supports mm, µm and nm which will always be offer even if the
  underlying scannable is in energy etc.

Change-Id: I4abb55d1e188cd53fb601839d360fdccca580c94