I20_1-95 : Add GUI for setting scannable/PV to be recorded during linear experiment
This commit adds GUI components to the 'Time resolved experiment' view
for setting up the list of scannables/PVs to be recorded at each point
during a linear experiment scan. A dialog box is used to setup the list,
which is then stored in TimeResolvedExperimentModel to be used when
running a scan. Current values are also stored in the preference store.
Also, Tfg triggered scans now also have the scannables/PVs added to them.

Change-Id: Ic79d28e9344d9686ea4738d1f8d7e7836a00eb25